requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.3.0
Pillow>=8.0.0
imageio>=2.9.0
scipy>=1.6.0
spectral>=0.22.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
joblib>=1.0.0
PyQt5>=5.15.0
PyYAML>=5.4.0
tqdm>=4.60.0
